+++ tests/tests_table.py
@@ -84,18 +84,117 @@ class TableTestCase(unittest.TestCase):
         self.assertIn('does not match format',
                       context_manager.exception.message)
 
-    def test_table_getitem_error(self):
-        with self.assertRaises(ValueError) as context_manager:
-            self.table['test']
+    def test_table_getitem_invalid_type(self):
+        with self.assertRaises(ValueError) as exception_context:
+            self.table[3.14]
+        self.assertEqual(exception_context.exception.message,
+                         'Unsupported key type: float')
 
-    def test_table_setitem(self):
+        with self.assertRaises(ValueError) as exception_context:
+            self.table[b'name']
+        self.assertEqual(exception_context.exception.message,
+                         'Unsupported key type: str')
+        # TODO: should change to 'bytes' on Python3
+
+    def test_table_getitem_column_doesnt_exist(self):
+        with self.assertRaises(KeyError) as exception_context:
+            self.table['doesnt-exist']
+
+        self.assertEqual(exception_context.exception.message,
+                         'doesnt-exist')
+
+    def test_table_getitem_column_happy_path(self):
+        expected_values = ['Álvaro Justen', 'Somebody', 'Douglas Adams']
+        self.assertEqual(self.table['name'], expected_values)
+
+        expected_values = [
+                datetime.date(1987, 4, 29),
+                datetime.date(1990, 2, 1),
+                datetime.date(1952, 3, 11)]
+        self.assertEqual(self.table['birthdate'], expected_values)
+
+    def test_table_setitem_row(self):
         self.first_row['name'] = 'turicas'
         self.first_row['birthdate'] = datetime.date(2000, 1, 1)
         self.table[0] = self.first_row
         self.assertEqual(self.table[0].name, 'turicas')
         self.assertEqual(self.table[0].birthdate, datetime.date(2000, 1, 1))
 
-    def test_table_delitem(self):
+    def test_field_names_and_types(self):
+        self.assertEqual(self.table.field_names, self.table.fields.keys())
+        self.assertEqual(self.table.field_types, self.table.fields.values())
+
+    def test_table_setitem_column_happy_path_new_column(self):
+        number_of_fields = len(self.table.fields)
+        self.assertEqual(len(self.table), 3)
+
+        self.table['user_id'] = [4, 5, 6]
+
+        self.assertEqual(len(self.table), 3)
+        self.assertEqual(len(self.table.fields), number_of_fields + 1)
+
+        self.assertIn('user_id', self.table.fields)
+        self.assertIs(self.table.fields['user_id'], rows.fields.IntegerField)
+        self.assertEqual(self.table[0].user_id, 4)
+        self.assertEqual(self.table[1].user_id, 5)
+        self.assertEqual(self.table[2].user_id, 6)
+
+    def test_table_setitem_column_happy_path_replace_column(self):
+        number_of_fields = len(self.table.fields)
+        self.assertEqual(len(self.table), 3)
+
+        self.table['name'] = [4, 5, 6]  # change values *and* type
+
+        self.assertEqual(len(self.table), 3)
+        self.assertEqual(len(self.table.fields), number_of_fields)
+
+        self.assertIn('name', self.table.fields)
+        self.assertIs(self.table.fields['name'], rows.fields.IntegerField)
+        self.assertEqual(self.table[0].name, 4)
+        self.assertEqual(self.table[1].name, 5)
+        self.assertEqual(self.table[2].name, 6)
+
+
+    def test_table_setitem_column_slug_field_name(self):
+        self.assertNotIn('user_id', self.table.fields)
+        self.table['User ID'] = [4, 5, 6]
+        self.assertIn('user_id', self.table.fields)
+
+    def test_table_setitem_column_invalid_length(self):
+        number_of_fields = len(self.table.fields)
+        self.assertEqual(len(self.table), 3)
+
+        with self.assertRaises(ValueError) as exception_context:
+            self.table['user_id'] = [4, 5]  # list len should be 3
+
+        self.assertEqual(len(self.table), 3)
+        self.assertEqual(len(self.table.fields), number_of_fields)
+        self.assertEqual(exception_context.exception.message,
+                         'Values length (2) should be the same as Table '
+                         'length (3)')
+
+    def test_table_setitem_invalid_type(self):
+        fields = self.table.fields.copy()
+        self.assertEqual(len(self.table), 3)
+
+        with self.assertRaises(ValueError) as exception_context:
+            self.table[3.14] = []
+
+        self.assertEqual(len(self.table), 3)  # should not add any row
+        self.assertDictEqual(fields, self.table.fields)  # should not add field
+        self.assertEqual(exception_context.exception.message,
+                         'Unsupported key type: float')
+
+        with self.assertRaises(ValueError) as exception_context:
+            self.table[b'some_value'] = []
+
+        self.assertEqual(len(self.table), 3)  # should not add any row
+        self.assertDictEqual(fields, self.table.fields)  # should not add field
+        self.assertEqual(exception_context.exception.message,
+                         'Unsupported key type: str')
+        # TODO: should change to 'bytes' on Python3
+
+    def test_table_delitem_row(self):
         table_rows = [row for row in self.table]
         before = len(self.table)
         del self.table[0]
@@ -103,6 +202,50 @@ class TableTestCase(unittest.TestCase):
         self.assertEqual(after, before - 1)
         for row, expected_row in zip(self.table, table_rows[1:]):
             self.assertEqual(row, expected_row)
+
+    def test_table_delitem_column_doesnt_exist(self):
+        with self.assertRaises(KeyError) as exception_context:
+            del self.table['doesnt-exist']
+
+        self.assertEqual(exception_context.exception.message,
+                         'doesnt-exist')
+
+    def test_table_delitem_column_happy_path(self):
+        fields = self.table.fields.copy()
+        self.assertEqual(len(self.table), 3)
+
+        del self.table['name']
+
+        self.assertEqual(len(self.table), 3)  # should not del any row
+        self.assertEqual(len(self.table.fields), len(fields) - 1)
+
+        self.assertDictEqual(dict(self.table[0]._asdict()),
+                             {'birthdate': datetime.date(1987, 4, 29)})
+        self.assertDictEqual(dict(self.table[1]._asdict()),
+                             {'birthdate': datetime.date(1990, 2, 1)})
+        self.assertDictEqual(dict(self.table[2]._asdict()),
+                             {'birthdate': datetime.date(1952, 3, 11)})
+
+    def test_table_delitem_column_invalid_type(self):
+        fields = self.table.fields.copy()
+        self.assertEqual(len(self.table), 3)
+
+        with self.assertRaises(ValueError) as exception_context:
+            del self.table[3.14]
+
+        self.assertEqual(len(self.table), 3)  # should not del any row
+        self.assertDictEqual(fields, self.table.fields)  # should not del field
+        self.assertEqual(exception_context.exception.message,
+                         'Unsupported key type: float')
+
+        with self.assertRaises(ValueError) as exception_context:
+            self.table[b'name'] = []  # u'name' actually exists
+
+        self.assertEqual(len(self.table), 3)  # should not del any row
+        self.assertDictEqual(fields, self.table.fields)  # should not del field
+        self.assertEqual(exception_context.exception.message,
+                         'Unsupported key type: str')
+        # TODO: should change to 'bytes' on Python3
 
     def test_table_add(self):
         self.assertIs(self.table + 0, self.table)